chore: rename to match py samples naming convention
