+++ genai/count_tokens/counttoken_compute_with_txt.go
@@ -15,8 +15,6 @@
 // Package count_tokens shows examples of counting tokens using the GenAI SDK.
 package count_tokens
 
-// TODO[vburlaka@google.com,msampathkumar@google.com]: Remove `count_tokens` region tags after Feb 2025
-// [START googlegenaisdk_count_tokens_compute_with_txt]
 // [START googlegenaisdk_counttoken_compute_with_txt]
 import (
 	"context"
@@ -91,4 +89,3 @@ func computeWithTxt(w io.Writer) error {
 }
 
 // [END googlegenaisdk_counttoken_compute_with_txt]
-// [END googlegenaisdk_count_tokens_compute_with_txt]
